feat: add nginx service with port mapping in validate_nginx job
The `validate_nginx` job in the workflow now includes a service called `backend` with the `nginx:1.25.4` image. It is mapped to port 4000 on the host machine. This allows the job to interact with the nginx service running in the container.

Additionally, the `docker run` command in the `validate_nginx` job has been updated to include the `-v` flag to mount the `nginx/nginx.conf` file as the nginx configuration file in the container. The nginx configuration is then tested using the `nginx -t` command.